core: set VmBackup toCheckpointId only after checkpoint created
The Engine shouldn't report on the created checkpoint ID
(to_checkpoint_id) before it was really created by the host.

We need to avoid the scenario that a backup application uses the
checkpoint ID even though that the checkpoint wasn't created yet,
this case can happen when there was a failure during the backup
process and the operation failed.

This patch adds another member to hold the checkpoint ID to the
parameters class, the parameters persisted in the DB for each change.
Only after the checkpoint created by the host, the VmBackup entity will
contain the created checkpoint ID.

Change-Id: I30ec4b6df1fad26c4f3affa34ba6a9cd92fc1a03
Signed-off-by: Eyal Shenitzky <eshenitz@redhat.com>