core: prepare and measure on the same host
An image has to be measured on the same host it is prepared
on, otherwise the LV may not be active on the host which
measures and it will fail with ENOENT.

This patch uses getVdsRunningOn passed on from the parent
command to execute the measurement.

The patch also changes runVdsCommandWithFailover to
runVdsCommandWithoutFailover as we cannot really safely
failover, since the LV may not be prepared on the new
host.

Change-Id: I7aee292d9134e9b33c23c69b0d3736e6321605c7
Bug-Url: https://bugzilla.redhat.com/1820182
Signed-off-by: Benny Zlotnik <bzlotnik@redhat.com>