restapi - GET vm-nic: don't include reported-devices
Currently when getting nics of a VM, the reported-devices
of the nics (when exist) are returned inside the Nic
representation. This is incorrect, as they are bound to
a different context in the API tree:

  GET .../api/vms/xxx/nics

  GET .../api/vms/xxx/nics/yyy/reporteddevices

The reported-devices won't be returned in the presence
of <All-Content=true> header either.

If a client wishes to get the reported devices inside
the nics as part of the same query, it should use the
link-following mechanism:

  GET .../api/vms/xxx/nics?follow=reporteddevices

Temporarily this usage won't work due to Bugzilla 1821624
(link not provided in order not to confuse automatic tools).
This bug is expected to be resolved shortly, but it's no
reason not to merge this patch.

This patch also eliminates getting statistics when
<All-Content=true> header exists, since this is wrong
behavior. Statistics may be fetched inline using:

   GET .../api/vms/xxx/nics?follow=statistics

This works and is unaffected by Bugzilla 1821624 which
applies only to two-word links (reported-devices)

Bug-Url: https://bugzilla.redhat.com/1723309
Change-Id: I2d328e8ff9ec23633f8ed11d5938359a2d5a0c69
Signed-off-by: Ori_Liel <oliel@redhat.com>