Disable iptables support on engine-setup
Prevent selecting iptables as firewall manager and remove the
deprecation notice.

User can still force using or asking by overriding this in an answer
file, or setup config file. No code is actually removed or disabled -
only a setting in setup conf tells that only firewalld is valid.

We added the deprecation notice in 4.2, intended to remove in 4.3 but
apparently forgot. Do this now, for 4.4.

Change-Id: I8a9613acbfcd75b4590cf77c8496f9f6ee105938
Signed-off-by: Yedidyah Bar David <didi@redhat.com>