API: change redefine checkpoint and delete checkpoint API
According to the changes in libvirt, in order to redefine a checkpoint
libvirt needs the checkpoint XML. The Engine stores this checkpoint XML
after checkpoint creation in its database.

The API to redefine the checkpoints of a VM changed to redefine a list
of checkpoint each time instead of redefining all the checkpoints in one
API call, this is needed because the checkpoint XML contains also the VM
domain XML and uses a big amount of data. Separating the redefinition
of the checkpoints will allow the Engine to know exactly which
checkpoint redefinition failed and maybe recover from it.

So now, the Engine request should be:

VM.redefine_checkpoints -
{
  'vmID': vm_id,
  'checkpoints': [
    {
     'id': 'checkpoint_1',
     'xml': 'checkpoint_1_xml'
    },
    {
     'id': 'checkpoint_2',
     'xml': 'checkpoint_2_xml'
    },
    ...
  ]
}

The response will be a list of checkpoint IDs that were redefined
successfully.
If checkpoint redefinition failed, the list will not include the failed
checkpoint ID and the ones that follow it.

For deleting checkpoints, the behavior will be similar but the list will
include only the checkpoints IDs:

VM.delete_checkpoint -
{
  'vmID': vm_id,
  'checkpoint_ids': [checkpoint_1, checkpoint_2, ... ]
}

The response will be a list of checkpoint IDs that were deleted
successfully.
If checkpoint deletion failed, the list will not include the failed
checkpoint ID and the ones that follow it.

The drawback of this change is that the operations will be longer because
now the Engine will redefine or delete the checkpoints in bulks.

Change-Id: I79ac69f9d580b57a37102e6c02c56727f2f79aa1
Signed-off-by: Eyal Shenitzky <eshenitz@redhat.com>