ansible: check image update status for node-ng
ovirt-node-ng is updated as an image using the imgbased utility that
runs inside a %post scriptlet of the image-update rpm.
If imgbased fails to upgrade the image, the %post scriplet would fail
but this will not cause dnf to fail.

imgbased touches a file in /var/imgbased/.image-updated on successful
update, so we can test if the file exist to make sure the upgrade was
ok, otherwise fail the task.

Change-Id: I27454becb06d1c2aee55b8e9ffe767e2e7ff5aed
Bug-Url: https://bugzilla.redhat.com/1770893
Signed-off-by: Yuval Turgeman <yturgema@redhat.com>