packaging: Add imageio setup plugin
Add new engine-setup plugin for setting up imageio.

Imageio has to be always installed and we don't provide any user
customization. We need to configure firewall and remote service
for imageio daemon. Imageio port is hardcoded and only things
which needs to be configured in daemon configuration is remote address
and paths to certificates used for TLS communication. We also need
to overwrite control service setup to use communication over TCP
port, but this is done in static configuration template and doesn't
need any environment specific setup.

We require ovirt-imageio-daemon 2.0.4, the first version compatible with
engine.

Change-Id: I0fd798b46500c594d9a6537f6a6cab519e597722
Bug-Url: https://bugzilla.redhat.com/1801710
Signed-off-by: Vojtech Juranek <vjuranek@redhat.com>