core: use the host a VM is running on to measure
MeasureVolume does not prepare the volumes if the volume
is attached to a running VM, since it is already prepared.
However, since the calling commands choose a host randomly,
the host selected for the measurement is not necessarily
the one running the VM, which will result in ENOENT.

This patch uses the host the VM is running on. Unless it is
not suitable to measure (older cluster level), then it will
fallback to the legacy measurement.

Change-Id: I31585929c8c947940cca48cd88f8b2fffe2c8a67
Bug-Url: https://bugzilla.redhat.com/1820182
Signed-off-by: Benny Zlotnik <bzlotnik@redhat.com>