webadmin: add 'no entity present' state to EntityModel
Enable EntityModel to be used as an option(maybe) container type.
From a logical point of view EntityModel can have 2 states now:
 1. 'entity present' (aka Some)
   - for backward compatibility it's default state
   - contains value or null
 2. 'no entity present' (aka None)
   - the  container is considered empty
   - due to backward compatibility the call to getEntity()
     will not throw NoSuchElementException but will return whatever
     value is currently stored (which should be ignored as it's stale)
   - empty container is not valid

Background
If editor is in invalid state then the current value in the model
should be removed. This is implemented as resetting the value to 'null'
and works well for required fields. However 'null' is a legal value
for optional fields and cannot be used as special 'no value' marker.
Using 'null' for such purpose resulted in  previous value
being cleared silently.

Change-Id: Ib1d352afc3e77073adab9fe2f2e6661e8229d568
Bug-Url: https://bugzilla.redhat.com/1455944
Signed-off-by: Radoslaw Szwajkowski <rszwajko@redhat.com>