setup: Update old vmconsole certificates
When a new Engine CA certificate is installed, vmconsole certificates
must be updated too.  The current condition in the vmconsole
certificate check looks only at presence of vmconsole stuff, not its
modification time.  If the certificates already exist, nothing is
updated and vmconsole stops to work.  This patch fixes the condition.

Change-Id: I4eda3333ef6fbb3acb63f4ff69f6dc9a71f7d2a0
Signed-off-by: Milan Zamazal <mzamazal@redhat.com>