core: Change initial default of cluster.bios_type
When the default value for cluster.bios_type column is set in
04_04_0150_add_bios_type_column_to_cluster.sql upgrade script (where the
column is initially created), all existing records in cluster table
receive this value. That value is not updated after the default is
changed later in 04_04_0580_change_cluster_bios_type_default.sql.

To fix this, the default value in
04_04_0150_add_bios_type_column_to_cluster.sql must be changed also.

Change-Id: Iae3e10a28322b8422e84ccae20e1f8dd0864d9c2
Allow-db-upgrade-script-changes: Yes
Bug-Url: https://bugzilla.redhat.com/1824472
Signed-off-by: Shmuel Melamud <smelamud@redhat.com>