core: Update branding module's dependencies
When trying to use the branding module with an add-on deployment
(i.e. a branding package or ovirt-web-ui), there are dependencies
issues.  The JSP and JSTL dependencies needed to be added.

Now, if an add-on wants to use branding, the ear/war can depend
on the "org.ovirt.engine.core.branding" without error.  To be
able to use the "obrand" JSP taglib, the branding.jar needs to be
copied/symlinked to the war's WEB-INF/lib.

Change-Id: I592dddbb3201f494472d6e93fa933318ec8cd4c7
Signed-off-by: Scott J Dickerson <sdickers@redhat.com>