webadmin: Close connection after upload
The UI did not close the connection after upload. This cause the proxy
to keep the connection open to the daemon, which keeps the backend open.
Since the backend is open, deactivating volume on block storage will
fail after upload.

This issue was hidden with the old proxy since it did not support
keep-alive connection, and closed the connection after every request.

A sane application control the HTTP connection but in browser
environment this is impossible, and we cannot even set the "Connection:
close" request header, since it is considered unsafe[1][2].

Fix by using new "close" query argument introduced by imageio 2.0.5.
With this fix the proxy will close the connection after the last
request:

    PUT /images/ticket-id?close=y

Closing the connection in the proxy will close the HTTP backend, which
will close the connection to the daemon, and the daemon will close the
backend.

[1] https://xhr.spec.whatwg.org/#the-setrequestheader()-method
[2] https://fetch.spec.whatwg.org/#forbidden-header-name

Change-Id: I6e550cf307d94f95ef7c4acf920cf625a0eba791
Signed-off-by: Nir Soffer <nsoffer@redhat.com>