webadmin: Close connection after download
When downloading images we have no control over connection handling.
The browser will try to keep the connection open to optimize additional
requests for this domain, but imageio proxy and daemon expect clients to
close the connection when done. Not closing the connection will fail
deactivation of the volume after the download because the daemon keeps
the file backend until the HTTP connection is closed.

In ovirt-imageio 2.0.5 we can use "close" query argument to close the
connection after the request.

Change-Id: I05288e4ced74a5050b533f969be2c6c87dc7d9b7
Signed-off-by: Nir Soffer <nsoffer@redhat.com>