core: snapshot: allow force freeze in engine
The patch will prevent the situation of unsuccessful freeze command
happening in VDSM and resulting in data corruption. When the freeze
command will reach the timeout in the VDS command, the snapshot won't
start. The current isSnapshotUsed check assumes the dumpxml will
return the chain with the snapshot. But, the chain hasn't been
updated. There is no reliable way to tell if the volume is used or
not from the engine.

This patch will allow the user to select in engine-config whether he
wishes to perform the freeze and thaw command in the engine instead
of doing so in VDSM. The value for it is:
'LiveSnapshotPerformFreezeInEngine', by default set to 'true'.
This way if freeze is stuck we can fail the snapshot and prevent
the above situation.

Change-Id: I47f1c38f4fe8d3215e8f8cbaa9835f5647df1ff7
Bug-Url: https://bugzilla.redhat.com/1821164
Bug-Url: https://bugzilla.redhat.com/1764943
Signed-off-by: Liran Rotenberg <lrotenbe@redhat.com>