core: snapshot: allow inconsistent snapshot
This patch will let snapshot to be executed even when failing to
freeze in the engine.

A new value in engine-config will let the user choose if he wishes to
continue doing snapshot even when freeze fails. This value is:
'LiveSnapshotAllowInconsistent', by default set to 'true'. This may
result in inconsistent snapshot and on the risk of the user.

Change-Id: Id8dfd47f9fe0c11459ace54c55b399a5a52fba5f
Bug-Url: https://bugzilla.redhat.com/1821164
Bug-Url: https://bugzilla.redhat.com/1764943
Signed-off-by: Liran Rotenberg <lrotenbe@redhat.com>