core: use Arrays.asList instead of List.of
List#of will create an immutable list and assign this type for
DiskImage, this may look harmless, but ImmutableCollections are
not serializable by Jackson as it does not have a public constructor.
In turn, it will break compensation and mess up the engine start-up.

This patch instead uses Arrays#asList which returns a
de-serialization-friendly ArrayList.

Change-Id: I70d9a7b31d0383ed1ad2040d62102ab63dfbef8e
Bug-Url: https://bugzilla.redhat.com/1827033
Signed-off-by: Benny Zlotnik <bzlotnik@redhat.com>