core: Replaced null ids with image ids
When importing a template from ovirt-engine-sdk
scripts, the StoragePoolId and the
SourceStorageDomainId are null. This causes
the template ova import to crash.

Checks if the StoragePoolId and SourceStorageDomainId
are null values and if so sets the appropriate values
from the Disk Image object so that the import ova
template completes successfully.

Change-Id: I0015cc591bc5022e024e24d759015f4cc457ae15
Bug-Url: https://bugzilla.redhat.com/1830762
Signed-off-by: Steven Rosenberg <srosenbe@redhat.com>