packaging: database: Fix createUser
createUser was added to allow standalone creation of a user, by
engine-backup. It used the existing function _performDatabase that tried
to also create a database if one was asked for. For engine-backup's
case, that was enough - because it either wanted to create db+user or
only user.

Now I want to use createUser for creating a read-only user on the DWH
database for grafana. We want this also on upgrade, where we want to
create only the user, but also already have the database.

Fix this by simply copying the parts we need from _performDatabase. It
might make sense to completely remove this function and split it to two
(one for db, one for user), but I do not need this right now and it's a
bit delicate.

Change-Id: Idc9fa980747a2cc588cd57813fac90d4b16f0fa3
Signed-off-by: Yedidyah Bar David <didi@redhat.com>