packaging: move hardcoded strings to encoded
Since now we in 4.4 the engine is python3, the strings are now using
bytes. The hardcoded string didn't change to bytes. Therefore, when
we tried to envoke comparison in the socket, we failed. This broke
noVNC console with encryption.

This patch will set the hardcoded string with bytes, there is no need
for back compatibility since 4.4 will run only on machines with el8
or newer, which are using python3.

Change-Id: I153935061ebfbedcaeb0e4f305b1c891d33d505a
Bug-Url: https://bugzilla.redhat.com/1826431
Signed-off-by: Liran Rotenberg <lrotenbe@redhat.com>