core: Fix notify_time parsing when value < MAX_INT
Depending on the uptime of a host, and the monotonic time source used by
vdsm for computing event "notify_time", the value may be integer or
long.

>>> int(time.monotonic() * 1000) > (2**31 - 1)
False
>>> int(os.times()[4] * 1000) > (2**31 - 1)
True

We assumed that notify_time is always long value, and returned null if
it was integer value. This results in VdsmVm with a null timestamp,
which fails VM monitoring and cause the VM to get stuck in the current
state.

2020-04-26 20:40:46,509+03 ERROR [org.ovirt.engine.core.vdsbroker.monitoring.VmsMonitoring]
(EE-ManagedScheduledExecutorService-engineScheduledThreadPool-Thread-46) [] Failed during
vms monitoring on host host2 error is: {}: java.lang.NullPointerException
    at deployment.engine.ear//org.ovirt.engine.core.vdsbroker.VmManager.isLatestData(VmManager.java:220)
    ...

Fixed by using Number instead of Long.

Change-Id: I6284afe99a0cfcee532f4d514159412feb235989
Bug-Url: https://bugzilla.redhat.com/1828088
Signed-off-by: Nir Soffer <nsoffer@redhat.com>