core: redefine VM checkpoints according to the VM defined checkpoints
Checkpoint redefinition is needed in case the VM powered-off or crashed
and in other failure flows cases when the checkpoints chain in the
Engine DB and the defined checkpoints chain on the VM that reported by
libvirt is not the same.
This patch adds the logic of how to handle checkpoint redefinition.

There are a few scenarios the Engine should handle -

 1. The VM crashed or powered-off - VM has no defined checkpoints and
    the Engine should define the entire chain

 2. Part of the checkpoints chain defined on the VM - Engine should
    define the missing checkpoints in the chain.

 3. There is a mismatch between the checkpoint chain in the Engine DB
    and the defined checkpoint on the VM - Engine should remove the
    chain from libvirt and from the DB.

This patch adds the logic of how to distinguish between the scenarios and
how to handle options 1 and 2, for the 3rd option, the Engine will log
the error and fail the backup operation, this option will be implemented
in the future.

Change-Id: Iaf47db0cb179db393edeb6e23d73aeccbab4dce5
Signed-off-by: Eyal Shenitzky <eshenitz@redhat.com>