core: Add TPM device
This patch adds support for emulated TPM devices in VMs.  We use
software emulated TPM 2.0 devices.  Those devices have the following
requirements:

- They work only with UEFI BIOS.
- They require swtpm on the hosts.

swtpm dependency is added to Vdsm in 4.4 thus we limit TPM support to
4.4+ clusters.  This is not a hard requirement, just a simplification.

TPM devices are enabled or disabled for each given VM.  The initial
value is taken from the VM template and the VM template default value
is false.  If the general requirements for TPM device (cluster level,
UEFI BIOS) are not satisfied, the option is not editable and is set to
false.

This functionality is not very useful without TPM data persistence.
The data persistence will be implemented in followup patches.
For this reason, TPM devices are disabled in config by default now.

Change-Id: I53069e9e4bed0bde1118e6e6066e287dd173534f
Signed-off-by: Milan Zamazal <mzamazal@redhat.com>