core: prevent from doing an incremental and full backup operation
Due to bz #1829829, Libvirt doesn't handle the case of mixing
full and incremental backup under the same operation.

This situation can happen in oVirt when adding a new disk
to a VM that already has a previous backup.
It will create a request from libvirt to start a backup that contains
disks that are not included in the previous checkpoint.

This patch blocks this operation and instructs the user to create
a full backup that includes all the disks.

Change-Id: I2ec7705d5d3c365941b2177da237a4e7ae21b074
Signed-off-by: Eyal Shenitzky <eshenitz@redhat.com>