core: correct volume size after creation
Due to extent alignment happening when creating new volumes[1],
specifically on block domains, engine can end up having the wrong size
stored in the database, the wrong size, when used later will cause
failures because snapshots will be created using the wrong size.

This patch fixes the issue by asking vdsm to report the size for the
image it created and updating the size in the engine to ensure both see
the same size.

[1] https://github.com/oVirt/vdsm/blob/master/lib/vdsm/storage/volume.py#L1181

Change-Id: I329c4d7d2fc7322e4165acaaa678c5b553c15e01
Bug-Url: https://bugzilla.redhat.com/1758048
Signed-off-by: Benny Zlotnik <bzlotnik@redhat.com>