packaging: engine-backup: Exclude scl conf files on restore
/etc/ovirt-engine/engine.conf.d/10-scl-postgres.conf and
/etc/ovirt-engine-dwh/ovirt-engine-dwhd.conf.d/10-scl-postgres.conf

were packaged in 4.3, in order to configure the version of SCL
PostgreSQL we should use.

On restore, in 4.4, we do not need these anymore. All the code we had
in the engine and in dwh that relied on $sclenv was already removed.
I didn't run into any issues while having them, but it's somewhat
misleading, and in theory can also be harmful - if there some code
somewhere that still depends on $sclenv, it will break.

So, do not restore them.

Change-Id: I2af772476c654195ab30edf10d6fcc326b5e0179
Signed-off-by: Yedidyah Bar David <didi@redhat.com>