core: fix unwanted vm status change
When we create a VM from snapshot we don't need the VM to be down.
Therefore it was wrong to change the status to ImageLocked when
starting the command and to down when finishing the command. While in
the host, the VM is running. In the wrong timing, it would let the
user to execute command on the VM as it's down.

This patch will remove the status changes for the
'AddVmTemplateFromSnapshot' command and save the VM from forbidden
actions.
The snapshot we are using will be locked during the command.

Change-Id: I3700f02773e2c160ff9b35917655c5e91f70fea0
Bug-Url: https://bugzilla.redhat.com/1849370
Signed-off-by: Liran Rotenberg <lrotenbe@redhat.com>