core: fix import template from ova via the sdk
When importing a template from configuration (not as part of
register template), StoragePoolId and the SourceStorageDomainId
are not set. This caused the import template from ova via the sdk
to fail when trying to retrieve the virtual size of its disks
through qemu-img.

This fixes this issue by setting the storage pool to the one the
template is added to and the storage domain of each disk is
queried from the database.

Change-Id: Ifdad5bd2a7d3e6ba3c064536e93932ba0c960fc5
Bug-Url: https://bugzilla.redhat.com/1830762
Co-authored-by: Arik Hadas <ahadas@redhat.com>
Signed-off-by: Steven Rosenberg <srosenbe@redhat.com>