core: read installed virtio-win agents
This patch will read the available VirtIO-Win agents. The agents
suppose to be installed by a separate RPM. We need to read what is
available in order to be able comparing them to the existing in the
windows guests.
From this data we need to pull only the guest-agent information.

The JSON that suppose to come with the RPM will look like:
{
  "agents": [
    {
      "arch": "amd64",
      "agent_version": "101.1.0-1",
      "name": "Red Hat QEMU guest agent",
    },
...
]}

The data will be read by the OS ID, taking care about the architecture
and the specific windows OS. The version of the QEMU guest agent will be
save within the class.

Change-Id: I437127fe2a6f5072353b609a13e877e744c6185e
Bug-Url: https://bugzilla.redhat.com/1828347
Signed-off-by: Liran Rotenberg <lrotenbe@redhat.com>