vdsbroker: fix duplicate disk name
Having multiple SCSI devices with host-passthough and custom property of
scsi_hostdev set to scsi_hd or scsi_block could cause a duplicate of the
disk name and therefore on the target device. This made a wrong domain
xml for a VM, making libvirt throwing exception when starting the VM.

This patch will take care of existing devices, skipping names in use.

Change-Id: Idd858cad969f459908f2e1b8300c9c334b77b6b2
Bug-Url: https://bugzilla.redhat.com/1834416
Signed-off-by: Liran Rotenberg <lrotenbe@redhat.com>