core: NVDIMM support
This patch adds support for NVDIMM devices.  From the point of view of
libvirt, NVDIMM devices are memory devices, but it's more practical to
treat them as host devices in oVirt.  NVDIMM devices are reported as
host devices from Vdsm, if the Vdsm version running on the host can
report them.

Since NVDIMM devices are memory devices, their size takes its part
from maxMemory.  Therefore we must increase maxMemory accordingly.
This part of memory is allocated as non-movable by default so it
shouldn't cause any problems.

Libvirt requires placing NVDIMM devices inside a NUMA node.  When no
NUMA is configured, NVDIMM devices can't be added.  The patch tries to
pick the right NUMA node for the device and refuses to add it if there
is no NUMA configured for the VM.

We set LABEL of a fixed size for each NVDIMM device.  Having a LABEL
is required on POWER and optional on x86_64.  The minimum LABEL size
as required by QEMU is 128 KiB.  The simplest to avoid LABEL presence
and size confusions is to always use LABEL of 128 KiB size.

The device listing in the Web UI doesn't list any NVDIMM device
properties besides the name and capability.  It might be useful to
display device mode and size but it's probably not worth of adding
another column to the listing.  Maybe device mode could be displayed
in `driver' column.

When a NVDIMM device is attached to a VM, memory snapshots are
forbidden.  This is to prevent inconsistencies between VM memory and
the NVDIMM memory.  Since validation from frontend with extended
database access (to check for NVDIMM devices) would be to cumbersome,
we make that check only in backend.

There is currently no support for hot plugging or migrating NVDIMM
devices in the platform.

Standard memory hot plug must be considered though.  DIMM hot plug is
possible in an el8 guest OS only when the hot plugged memory is
aligned on memory hot plug size.  libvirt selects the hot plugged DIMM
address dependent on the NVDIMM sizes.  The NVDIMM memory layout is
computed by the following QEMU code
(https://github.com/qemu/qemu/blob/v5.0.0/hw/mem/nvdimm.c#L143):

  pmem_size = size - nvdimm->label_size;
  nvdimm->label_data = memory_region_get_ram_ptr(mr) + pmem_size;
  pmem_size = QEMU_ALIGN_DOWN(pmem_size, align);

We must select the NVDIMM size to a value that ensures the proper
alignment in the end result.

Change-Id: Ie127051ca1fc230eb8077612e71db725690ac21c
Signed-off-by: Milan Zamazal <mzamazal@redhat.com>