core: add a foreign key to remove backup and checkpoint disks map
Add foreign key constrains to vm_backups and vm_checkpoints tables to
delete the disks map of the deleted backup / checkpoint.

There is no use-case of leaving the backup / checkpoint disks map
without the entity itself, the foreign key will save the need of calling
explicitly to the stored procedure of removing the disks map and will
make sure we aren't leaving the DB un-clean.

Change-Id: I42f479ccdb4e52bd2b612a25f788d917da67f89d
Signed-off-by: Eyal Shenitzky <eshenitz@redhat.com>