webadmin: show all disks when download denied
Downloading multiple disks for a running VM
is not allowed. At the current implementation
a failure popup will show only one VM and
one disk.

This patch will list the VMs and the Disks
that can't be downloaded including the reason,
for example: plugged to a running VM.

Screenshots: https://imgur.com/a/fczoMSs

Change-Id: I2ec83197d2c5f8ade0634295114180b3d7981d5b
Bug-Url: https://bugzilla.redhat.com/1662733
Signed-off-by: Ahmad Khiet <akhiet@redhat.com>