webadmin: Update error message
The CLUSTER_CANNOT_UPDATE_CLUSTER_FAILED_TO_UPDATE_VMS
error message is no longer relevant just for the
compatibility version change, but also for CPU and
BIOS type changes.

The error message is updated to be more general and the
user is asked to look in engine.log for details.

Change-Id: Ib9207b291b8526ff396540b34f7d9e03d91b3b86
Bug-Url: https://bugzilla.redhat.com/1830872
Signed-off-by: Lucia Jelinkova <ljelinko@redhat.com>