engine: Do not select default CPU type if empty
When changing the compatibility version of a cluster, the CPU type is
auto detected based on the CPU flags. This is done to detect the best
CPU type in case the old one was deprecated. However, when the CPU type
is empty (meaning that it should be auto detected when the first host
comes up), we should not change it.

Change-Id: I640e10cd54dc5306e5a8c02f8cbf1da40220a8a6
Signed-off-by: Lucia Jelinkova <ljelinko@redhat.com>