packaging: snapshot move freeze in engine to false
After discussion, doing freeze and thaw in the engine to live snapshot
command should be on the user responsibility and should be false by
default. Setting it to true might cause a VM having the file system
frozen without thawing.

This patch will change the value of LiveSnapshotPerformFreezeInEngine to
'false'.

Change-Id: Idaf8ec7c58721dcd845f88ae0994130dbd5a9299
Bug-Url: https://bugzilla.redhat.com/1837994
Signed-off-by: Liran Rotenberg <lrotenbe@redhat.com>