webadmin: Change Guest CPU value for HE
The information about the guest CPU was misleading for the hosted
engine. The VM is not managed by the engine and the cpu information is
not stored in the vm_dynamic table but we displayed the cluster value
which could or could not be actual.

For the same reason it is not possible to notify user that the VM uses
different CPU verb than is configured on cluster.

Change-Id: I0d666498f687d5e7686053964849a0ea1e50d110
Bug-Url: https://bugzilla.redhat.com/1830872
Signed-off-by: Lucia Jelinkova <ljelinko@redhat.com>