WebAdmin: Overriding incorrect Bios Type
After adding a Host to a legacy cluster, the
autodetect sets the cluster's bios type to
the legacy default. When editing the cluster,
the Bios Type is reset to the cluster default.
This prevents a VM from being created because
the cluster default does not have a valid Bios
Type.

Modified the code so that the cluster default would
only be set when the bios type is null. This preserves
the legacy bios type which allows for the proper chip
set to be obtained when a VM is created.

Change-Id: I24398a13af462b35b4099e0c93eb0f01b642f1ec
Bug-Url: https://bugzilla.redhat.com/1838439
Signed-off-by: Steven Rosenberg <srosenbe@redhat.com>