engine: Smartcard not editable for running VM
Editing smartcard VM field resulted in a bug where the value
was stored into the DB but the device was not added. Restarting the VM
did not help - user had to stop the VM, disable smartcard, enable
smartcard and run the VM.

Now the smartcard is not editable when the VM is running, but a next run
configuration is created. The changes take effect when the VM is down.

Change-Id: I89581f355b8759d29fc1451868183317cca54a3a
Bug-Url: https://bugzilla.redhat.com/1834523
Signed-off-by: Lucia Jelinkova <ljelinko@redhat.com>