core: fix snapshot engine freeze
When a VM snapshot contains memory, there's no risk of breaking data
consistency (i.e., no need for fs freeze/quiesce). Before this patch,
we mistakenly freezed the FS also before taking the snapshot that
contains memory, making the VM to be restored from the snapshot with a
frozen FS. That is avoided now by not freezing the FS in case the
snapshot contains memory.

Change-Id: I55a1f0837871cc7e81ae151ae4394dafcbe4d1e2
Bug-Url: https://bugzilla.redhat.com/1838493
Signed-off-by: Liran Rotenberg <lrotenbe@redhat.com>