packaging: setup: remote_engine: Various fixes
- Add dialog_name param
"manual_files" had it, but it was never passed to it. Add everywhere and
supply defaults. This is now more important than in the past, because we
now generate answerfiles using this name.

- Log stdout/stderr
Fix the collection logic, it was broken. We didn't ever notice, because
we didn't log.

- Fix public key verification
When enrolling a cert on a remote engine, we verify that the public key
as extracted from the keypair/csr we created is identical with the one
extracted from the cert created by the remote engine. For some reason,
in older versions of m2crypto/openssl, both sides were broken, but still
matched. Fix both to correctly extract the public key.

- Write pki artifacts as binary
Otherwise, the files contain stuff like:
b'-----BEGIN...
...
...-----\n'

Bug-Url: https://bugzilla.redhat.com/1624732
Change-Id: I3cc5b8696fff503effad9a54b7e682c865693a6b
Signed-off-by: Yedidyah Bar David <didi@redhat.com>