core: Addition of IPV6 hosts to hyperconverged cluster fails
when adding a host which is already part of cluster when we try
to re-add it to same cluster engine throws errors as
server already part of another cluster.

The validate condtion to check wheather the server is part of cluster
is checking for addr column in vds_interface_view which will have
only for ipv4 address. When ipv6 host is being added we need to
check with ipv6 address colum in the view.

Bug-Url: https://bugzilla.redhat.com/show_bug.cgi?id=1842004

Change-Id: Ie35fad6c4ca47639712cedec1bec8de520bd35d4
Signed-off-by: Ritesh Chikatwar <rchikatw@redhat.com>