core: PPC handle a console on HP VM
When the user set the VM via API to High Performance profile the VM
stays with VNC console. If he didn't remove it, in PPC architecture the
VM won't be able to boot.
The user, can set High Performance profile with console if he wishs to,
for that we need a USB controller.

This patch will examine the VM console in order to know if we can remove
the USB controller and input devices.

Change-Id: Ia9f749b9de8b6f27e4269c1647fc08b52dcc19cc
Bug-Url: https://bugzilla.redhat.com/1839398
Signed-off-by: Liran Rotenberg <lrotenbe@redhat.com>