core: clear compensation data after deactivating SD
When putting an SD into maintenance and 'Ignore OVF update failure'
is checked, and in case an error occurred during the OVF update,
the storage domain should enter to maintenance mode, but its state
transitions into 'Inactive' instead.
As part of the OVF update failure, the
DeactivateStorageDomainWithOvfUpdate command reaches to
endWithFailure() and executes a compensation, which changes the SD
status from 'preparing for maintenance' to 'unknown'.
To overcome that bug, that compensation data should be cleared because
endWithFailure is a part of the "positive" flow, after a successful
deactivation.

Change-Id: I3ecd114d8a9ab11dc71a4030252d6287cdd060e3
Bug-Url: https://bugzilla.redhat.com/1780910
Signed-off-by: Shani Leviim <sleviim@redhat.com>