core: Fixed upload ova template script error
When importing a template from the sdk example
script upload_ova_as_template.py the insertion
of the disk does not use the image id of the
ovf, but generates a new guid for the image id.
This allows for the ova file to be reused.

The issue with this script is that when inserting
the template after adding the disk, the engine
uses the image id of the ovf file instead of the
image id that was used to create the disk. This
causes a failure when adding the template when the
image id is used to validate the image size.

To fix this issue, the image id of the disk is
saved during initialization so that if the image
id of the ovf is different from the image id from
the database, the image id from the database that
correlates with the disk shall be used.

This prevents the failure reported where the image id
of the vdsm fails.

Change-Id: I1dd71e786f476e64b6089c662f90f7e5d5a49d53
Bug-Url: https://bugzilla.redhat.com/1819960
Signed-off-by: Steven Rosenberg <srosenbe@redhat.com>