search: cache invalid queries
In case that the query is invalid or erroneous or an exception is thrown,
the query is not cached, this makes the query parsed again and again
if the query origin comes from a scheduled timer.

This patch caches the problematic  query, so a null result will be returned
immediately to the caller after the first time, in addition, the log
message is written only once and not every time this query is
running which may load the CPU fill engine log with same repeating message.

Change-Id: I545f138160daabdd9f2831f3db9c1c5674a295d5
Bug-Url: https://bugzilla.redhat.com/1842495
Signed-off-by: Eli Mesika <emesika@redhat.com>