WebAdmin: Move the Remove VM button to the menu
Users can sometime be confused between the button
to remove the VM within a detailed tab view, and
the remove button for removing an item with the
tab detailed view. Such as when viewing a snapshot
within the snapshot tab of a VM within the Virtual
Machine screen. This means that the user can delete
the VM itself when their motives are to just say
remove a snapshot.

To make the system more user friendly, the remove
button has been moved to the kebana drop down menu
so that this mistake is mitigated. This change should
allow for greater usability.

Change-Id: I0ea5e626f054299b3973caf7d5860b46966bd6e9
Bug-Url: https://bugzilla.redhat.com/1763812
Signed-off-by: Steven Rosenberg <srosenbe@redhat.com>