engine: MDevicePolicyUnit considers nodisplay
When configuring mdev_type custom property for the VM, it is possible to
add string 'nodisplay' as the first value (before all mdev types). This
indicates that attribute display="on" should not be sent to libvirt.

The MDevicePolicyUnit did not consider this option and failed if it was
provided because it searched for it in in the mdev types reported by
the host.

This patch introduces MDevTypesUtils that is used by MDevicePolicyUnit
and LibvirtVmXmlBuilder. This should help to keep both classes in better
sync in the future.

Change-Id: Ifd9dee91526ce595f15098c1f6bef69046e9c0bc
Bug-Url: https://bugzilla.redhat.com/1844270
Signed-off-by: Lucia Jelinkova <ljelinko@redhat.com>