Core: Exporting a VM with disk as an OVA is RAW
When exporting a VM or template with a disk as an OVA,
the ovf file's metadata has the format of the disk set
to RAW even though the actual file format of the disk
is COW.

Overrode the cache image so that the format is cache
for processing.

This fixes the format issue.

Change-Id: I7bae1e8f6619bdb24c1cae4f802848b720205a59
Bug-Url: https://bugzilla.redhat.com/1845473
Signed-off-by: Steven Rosenberg <srosenbe@redhat.com>