webadmin: host interface list - unmanaged networks indicator
Display 'unmanaged' indicators on the interface list of a host. An
indicator  is displayed when a network exists only on the NIC of a host
and does not exist in the cluster. This indication is identical in terms
to the one in the setup networks dialog.
The indication is consistent with other indications on the interface
list:
- icon on the expansion link of the interface details view.
- column displaying indication per network in the interface network
  list.

Preview: https://imgur.com/a/G8G412V

Change-Id: Ie15bfaf4a48c2688b538e59a6098741d519810ea
Bug-Url: https://bugzilla.redhat.com/1477049
Signed-off-by: Eitan Raviv <eraviv@redhat.com>