core: Workaround to not have current user for scheduled jobs
Current user is populated only from logged-in user who executed
a command from UI or RESTAPI. But for methods executed from
a scheduler we don't have any logged-in user. This patch adds a code,
which assumes that null value in current user means that the method is
executed from scheduler and as such current user needs to be SuperUser.

Change-Id: Ifbb2e9dc3b41b276cb92fe77cd5e1d17b264d88e
Bug-Url: https://bugzilla.redhat.com/1846331
Signed-off-by: Martin Perina <mperina@redhat.com>