sso: fix userinfo response with external sso
This bug was introduced when support for plain text userinfo reponse had
been added. There is a legit flow when client_id is not available in sso session.
It was resolved to null causing internal NPE.

The solution is to default to encoded response when client_id is not
present because we are no able to determine the need of encoding.
This flow happens with external sso provider configured (ie. Keycloak).

Change-Id: I2191230e57cf7e0d2bb164bb350a67ad2e82cc1a
Bug-Url: https://bugzilla.redhat.com/1825907
Signed-off-by: Artur Socha <asocha@redhat.com>