WebAdmin: Change label to Free Huge Pages
Users complained that the Huge Pages label in the
Host's General Tab details screen was confusing and
that it was not clear which Huge Page values it
referred to.

Changing the label to "Free Huge Pages (size: amount):"
should make it clear that the label is for the free
Huge pages.

Change-Id: Iaf85ee18e55839e779ffcf6bc759afb37302e96d
Bug-Url: https://bugzilla.redhat.com/1806339
Signed-off-by: Steven Rosenberg <srosenbe@redhat.com>